Implementing Security Best Practices in Django

+++ advanced_features_and_security/LibraryProject/bookshelf/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import permission_required
 from .models import Book
+from .forms import BookForm  # Use a ModelForm for safe input handling
 
 
 # View all books (requires can_view)
@@ -14,17 +15,15 @@ def book_list(request):
 @permission_required("bookshelf.can_create", raise_exception=True)
 def book_create(request):
     if request.method == "POST":
-        title = request.POST.get("title")
-        author = request.POST.get("author")
-        year = request.POST.get("year")
-        Book.objects.create(
-            title=title,
-            author=author,
-            publication_year=year,
-            user=request.user
-        )
-        return redirect("book_list")
-    return render(request, "bookshelf/book_form.html")
+        form = BookForm(request.POST)
+        if form.is_valid():
+            book = form.save(commit=False)
+            book.user = request.user
+            book.save()
+            return redirect("book_list")
+    else:
+        form = BookForm()
+    return render(request, "bookshelf/book_form.html", {"form": form})
 
 
 # Edit a book (requires can_edit)
@@ -32,17 +31,21 @@ def book_create(request):
 def book_edit(request, pk):
     book = get_object_or_404(Book, pk=pk)
     if request.method == "POST":
-        book.title = request.POST.get("title")
-        book.author = request.POST.get("author")
-        book.publication_year = request.POST.get("year")
-        book.save()
-        return redirect("book_list")
-    return render(request, "bookshelf/book_form.html", {"book": book})
+        form = BookForm(request.POST, instance=book)
+        if form.is_valid():
+            form.save()
+            return redirect("book_list")
+    else:
+        form = BookForm(instance=book)
+    return render(request, "bookshelf/book_form.html", {"form": form, "book": book})
 
 
 # Delete a book (requires can_delete)
 @permission_required("bookshelf.can_delete", raise_exception=True)
 def book_delete(request, pk):
     book = get_object_or_404(Book, pk=pk)
-    book.delete()
-    return redirect("book_list")
+    if request.method == "POST":
+        book.delete()
+        return redirect("book_list")
+    return render(request, "bookshelf/book_confirm_delete.html", {"book": book})
+
